refactory remove opcontext
